+++ hass_cleaner/hass_cleaner/__init__.py
@@ -1,3 +1,3 @@
 """Hass-Cleaner Home Assistant App."""
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"